Do not destroy playlist when memory usage is high
The OS sends a memory warning to the plugin when device memory usage is
high, but there is no need to destroy the playlist, which should be able
to survive high-memory culling.

Other music-playing devices such as Spotify and SoundCloud do not halt
playback under similar observed conditions (i.e. when the device is
busy) - a playlist player is designed to work in the background, so
should not be prevented from doing so.